fix(#1472): CI 수정 — SonarCloud new code duplication 5.0% → 0%
- scripts/lib/stationNameIndex.js 신규 — normalizeStationName/buildNameIndex/lookupStationId
- scripts/build-station-distances.js는 helper들을 inline 정의 대신 require
- fetch-station-distances.js / fetch-station-travel-times.js와의 17줄 dup block 제거
  (fetch-* 두 파일은 별 개 PR로 점진적으로 옮길 예정 — surgical 룰 준수)
